Move select options into slug migration helper
Select options are a view responsibility
