+++ app/views/slug_migrations/edit.html.erb
@@ -8,7 +8,7 @@
       <% end %>
       <%=
         f.select :redirect_to,
-          grouped_options_for_select(@select_options),
+          grouped_options_for_select(select_options),
           {include_blank: "Please choose a path to redirect to"},
           {class: "select2 input-md-12 form-control"}
       %>
